add home screen and navigation

+++ background.sh
@@ -1,13 +1,14 @@
 start_client(){
-	cd client;
-	npm run metro &;
-	npm run emulator > emulator.log &;
+	cd client || exit;
+	npm run metro &
+	npm run emulator > emulator.log &
 	{
 		IFS= read -r pid;
-		booted=$(grep -q -m1 -oF 'boot completed');
-		kill -- "$pid";
+		if grep -q -m1 -oF 'boot completed'; then
+		  kill -- "$pid";
+		fi
 	} < <(echo "$BASHPID"; exec tail -n +1 -f emulator.log);
-	npm run android > app.log;
+	wait npm run android > app.log;
 	cd ..;
 	return 0;
 }
